i18n: English and Chinese Ui translatiton and switch actions

+++ src/ViewModel/MainWindow.py
@@ -1,11 +1,13 @@
 import os
 import shutil
+import sys
+from pathlib import Path
 
 from loguru import logger as log
 from PySide6 import QtCore
-from PySide6.QtCore import QSettings
-from PySide6.QtGui import QCloseEvent
-from PySide6.QtWidgets import QMessageBox
+from PySide6.QtCore import QProcess, QSettings
+from PySide6.QtGui import QAction, QCloseEvent
+from PySide6.QtWidgets import QApplication, QMessageBox
 
 from src import __version__
 from src.Model.Data.settings import (
@@ -80,11 +82,13 @@ class MainWindow_VM(MainWindow_View):
             self.guiSettings,
             "mainWindow_always_on_top",
         )
+
+        # Actions
         self.action_locate.triggered.connect(self.dialog_locate.show)
         self.action_preview.triggered.connect(self.tab_preview.preview_region)
         self.action_stitch.triggered.connect(self.tab_stitch.start_stitch)
         self.action_reclip.triggered.connect(self.tab_reclip.start_reclip)
-        self.action_print_score.triggered.connect(self.tab_reclip.printing_score)
+        self.action_print_score.triggered.connect(self.tab_reclip.printing_preview)
         self.action_output_pdf.triggered.connect(self.tab_reclip.save_pdf)
         self.action_select_folder.triggered.connect(
             self.tab_settings.select_score_working_folder
@@ -96,6 +100,7 @@ class MainWindow_VM(MainWindow_View):
         )
         self.action_rename_folder.triggered.connect(self.tab_settings.rename_folder)
         self.action_about.triggered.connect(lambda: About(self, __version__))
+        self.add_language_switch()
 
         # state bar
         self.label_version.setText("V" + __version__)
@@ -113,6 +118,50 @@ class MainWindow_VM(MainWindow_View):
         self.load_dock_perspective()
         self.toolBar_path.setFocus()
         self.appSettings.notice_all_observers()
+
+    def add_language_switch(self) -> None:
+        from src.View.MainWindow import LANGUAGES
+
+        actions: list[QAction] = []
+
+        def handel(action: QAction):
+            for a in actions:
+                a.blockSignals(True)
+                a.setEnabled(True)
+                a.setChecked(False)
+                a.blockSignals(False)
+            action.blockSignals(True)
+            action.setChecked(True)
+            action.setEnabled(False)
+            action.blockSignals(False)
+            lang = LANGUAGES[action.text()]
+            self.guiSettings.language = lang
+
+            # restart
+            self.close()
+            p = QProcess
+            app = QApplication.instance()
+            if not app:
+                return
+            exe = app.applicationFilePath()
+            if Path(exe).name.find("python") >= 0:
+                p.startDetached(
+                    sys.executable,
+                    [(Path(__file__).parent.parent.parent / "main.py").as_posix()],
+                )
+            else:
+                p.startDetached(exe)
+
+        for k, v in LANGUAGES.items():
+            action = QAction(self)
+            action.setText(k)
+            action.setCheckable(True)
+            if self.guiSettings.language == v:
+                action.setChecked(True)
+                action.setEnabled(False)
+            action.toggled.connect(lambda state, a=action: handel(a) if state else None)
+            self.menuLanguage.addAction(action)
+            actions.append(action)
 
     def load_dock_perspective(self) -> None:
         self.dock_manager.loadPerspectives(self.dockSettings)
@@ -143,7 +192,9 @@ class MainWindow_VM(MainWindow_View):
             return
 
         if self.captureThread:
-            log.warning("当前截图任务仍未结束，请稍后重试")
+            log.warning(
+                self.tr("Current capture task is not finished. Please try again later.")
+            )
             self.action_capture.setChecked(False)
             return
 
@@ -153,24 +204,38 @@ class MainWindow_VM(MainWindow_View):
             )
             messagebox = QMessageBox()
             messagebox.setWindowTitle(
-                f"工作目录下已存在{self.pathSettings.score_title}文件夹"
+                self.tr("Working folder is not empty: {}").format(
+                    self.pathSettings.score_title
+                )
             )
-            messagebox.setText(f"清空文件夹，或修改曲谱标题，并新建文件夹{new_title}")
-            messagebox.addButton("清空文件夹", QMessageBox.ButtonRole.YesRole)
-            messagebox.addButton("新建文件夹", QMessageBox.ButtonRole.NoRole)
-            messagebox.addButton("取消", QMessageBox.ButtonRole.NoRole)
+            text_clear_folder = self.tr("Clear folder")
+            text_new_folder = self.tr("New folder")
+            messagebox.setText(
+                self.tr(
+                    "Clear current folder, or switch to a new folder named: {}"
+                ).format(new_title)
+            )
+            messagebox.addButton(text_clear_folder, QMessageBox.ButtonRole.YesRole)
+            messagebox.addButton(text_new_folder, QMessageBox.ButtonRole.NoRole)
+            messagebox.addButton(self.tr("Cancel"), QMessageBox.ButtonRole.NoRole)
             messagebox.setWindowFlag(
                 QtCore.Qt.WindowType.WindowStaysOnTopHint, True
             )  # 设为置顶，必要
             messagebox.exec()
-            match messagebox.clickedButton().text():
-                case "清空文件夹":
-                    os.chdir(self.pathSettings.main_out_dir)  # release old dir
-                    shutil.rmtree(self.pathSettings.working_dir)
-                    log.success(f"已清空文件夹{self.pathSettings.score_title}")
-                case "新建文件夹":
-                    self.pathSettings.score_title = new_title
-                    log.success(f"已切换到新文件夹{self.pathSettings.score_title}")
+            result_text = messagebox.clickedButton().text()
+            if result_text == text_clear_folder:
+                os.chdir(self.pathSettings.main_out_dir)  # release old dir
+                shutil.rmtree(self.pathSettings.working_dir)
+                log.success(
+                    self.tr("Folder cleared: {}").format(self.pathSettings.score_title)
+                )
+            elif result_text == text_new_folder:
+                self.pathSettings.score_title = new_title
+                log.success(
+                    self.tr("Switched to new folder: {}").format(
+                        self.pathSettings.score_title
+                    )
+                )
             self.action_capture.setChecked(False)
             return  # break out
